+++ lib/rubocop.rb
@@ -60,6 +60,7 @@ require 'rubocop/cop/method_length'
 require 'rubocop/cop/op_method'
 require 'rubocop/cop/reduce_arguments'
 require 'rubocop/cop/percent_r'
+require 'rubocop/cop/favor_percent_r'
 
 require 'rubocop/report/report'
 require 'rubocop/report/plain_text'